requirements.txt
argparse
librosa>=0.8.0
pyyaml
scipy
soundfile
torch
wandb
python-dotenv
